refactor(migrations): refactor drizzle migrations script and updates gig-db and private-gig-db migrations
